requirements.txt
fastapi
uvicorn
python-dotenv
yt-dlp
spotipy
google-generativeai
requests
pydantic
ffmpeg-python
python-multipart
shazamio
audioop-lts; python_version >= "3.13"
static-ffmpeg